refactor: 활동 생성 API 명칭 정리 및 QR 스캐너 ID 전달 구조 개선
