+++ src/api/activity/post-activities.ts
@@ -1,16 +1,16 @@
 import { requestApi } from './request';
 import type { ApiResult } from './types';
 
-export interface MemberActivityRequest {
+export interface CreateActivityRequest {
   name: string;
   pointAmount: number;
 }
 
-export async function PostMemberActivities(
+export async function createActivity(
   name: string,
   pointAmount: number,
 ): Promise<ApiResult<unknown>> {
-  const requestBody: MemberActivityRequest = {
+  const requestBody: CreateActivityRequest = {
     name,
     pointAmount,
   };
